requirements.txt
playwright==1.48.0
python-dotenv==1.0.0
flask==3.0.0
werkzeug==3.0.1
gunicorn==21.2.0
